default network per tenant: - allow default network from object model - assign unassign default network as networks come and go - allow support for picking default network when unspecified - add unit and system tests

+++ Godeps/_workspace/src/github.com/contiv/objmodel/contivModel/cmExample/apiController.go
@@ -303,7 +303,7 @@ func (self *ApiController) ServiceCreate(service *contivModel.Service) error {
 	}
 
 	// Create service instances
-	for idx := int64(0); idx < service.Scale; idx++ {
+	for idx := 0; idx < service.Scale; idx++ {
 		instId := fmt.Sprintf("%d", idx+1)
 		var volumes []string
